[turbopack] Don't execute inventory_submit in rust analyzer (#83447)
### What?
In #83074 we changed how turbo task objects were registered to use the `inventory` crate.
This PR adds a new `inventory_submit!` macro that wraps the existing `inventory::submit` function with a `#[cfg(not(rust_analyzer))]` attribute to prevent Rust Analyzer choking on the generated code.

See example error

![image.png](https://app.graphite.dev/user-attachments/assets/2fe87f23-8717-4615-a86d-94c1b19ae099.png)

Because the inventory code does not affect any API surfaces this is safe and if anything should improve rust-analyzer performance