[turbopack] Attempt #2 at making rust-analyzer happy with our macros (#83461)
We are experiencing https://github.com/rust-lang/rust-analyzer/issues/19993 and Rust Analyzer is complaining about our macro gencode.  After some trial and error i determined the issue was our use of the `concat!` and specifically the inclusion of the crate name macro so this works around that.

Apparently Rust Analyzers implementation of concat! or maybe token tree parsing can get confused by string literals with `-` characters in them.  Simply avoiding that for rust analyzer is a reasonable short term workaround