Turbopack: Don't resolve tsconfig relative to file (#83484)
## What?

Changes in #83331 accidentally changed behavior of how `tsconfig.json`
is resolved. When using Next.js we don't want it to be resolved relative
to the current file being compiled. It should only use the application
root.

This PR fixes the behavior to match what it was before and adds a test
that fails if it breaks again.