Turbopack: Remove 20ms overhead on bootup (#83554)
## What?

This fetch, even though not awaited, was kicked off eagerly and the
first `fetch()` in Node.js loads the SSL certificate store which takes
roughly 20ms. Made sure that it's only fetched when the version is
needed, which is the first time errors happen, and then cached from that
point.


### Additional

In `test/integration/config-output-export/test/index.test.ts` nearly all
tests are flaky because of the way they're structured to do asserts
after finally that kills the application.

This causes it to race: Either the assertion is correct before it quits
or the assertion fails unexpectedly (flaking).

This PR fixes the test suite. Accidentally found this flakiness when I
made the change in this PR that makes the version step take more than 50
milliseconds the first time. That caused the assertion to be
consistently slower, thus reproducing the issue all the time.