Development: Only require loadJsConfig when using webpack (#83623)
## What?

Don't require/call loadJsConfig when using Turbopack, as it's handled
internally. loadJsConfig is one of the places where TypeScript gets
loaded. There's still the TS verify as well, this narrows the amount of
places where TypeScript gets loaded to make it easier to remove that
loading of TypeScript in a future PR.