Turbopack: remove `css_environment` from `Environment` (#83487)
Change the fix from #83334 to not have a "CSS environment" inside of every environment. I don't think that this makes sense, and it also makes the implementation more complicated.
```diff
  pub struct Environment {
    execution: ExecutionEnvironment,
+   css_environment: ResolvedVc<BrowserEnvironment>,
  }
```

Instead, we can just pass the client's `Environment` when we create the server's transform rules, to use the client's browserslist config for styled-jsx.

#83334 was also missing a test case

I recommend the review each commits individually.