tests: disable flaky deployment test while investigating upstream (#83705)
This test has started failing when deployed to hosted environments, but
it consistently is passing in `next start` tests.

We could retry the on-demand revalidation and the test will pass, but we
lose the value in the test.

This instead temporarily disables it until it can be investigated
upstream.

[slack
x-ref](https://vercel.slack.com/archives/C04KC8A53T7/p1757108995256419)