[Turbopack] Remove `ssr-data` module context (#83626)
SSR data is a special module context for development usecases to help with HMR of pages router pages.  This allows us to transform pages to strip the Component export, then the dev server can tell the difference between server side and client side changes which informs what kind of refresh we need to do.

Instead, create a new `ReferenceEntrySubType` and configure the `export stripping` transform to trigger using that.  This is a better approach since we only need to transform that module instead of putting a lot of dependencies in the new module-context and assert layer.

This should be a development time performance improvement for pages router

Finally delete a number of the `_runtime_entries` functions for server contexts.  These were always empty and were a user of the `ServerContextType::PagesData` enum varient that was being deleted.  

Closes PACK-4155