CI: use KV for test timing data (#83745)
Next.js tests are grouped based on timing data to better distribute them
for parallelization.

Previously this relied on a gist file that we would write to in CI.
However, this has more recently started resulting in 401s.

The actual reason for the 401s is the Turbo task wasn't propagating the
required environment variable to the task. It was mostly working by
accident, because we would store timing data on disk for the runner,
which would have been satisfied by a job that ran before it.

However despite that, relying on a gist file to read/write timing data
didn't feel like the write abstraction. I decided to refactor it to use
KV instead.

This also makes the handling more resilient to missing timing data. We
shouldn't fail the build if we can't fetch timings, as there's already
handling to fallback to round robin.