Turbopack: Remove matchers.reload() call on each request (#83720)
## What?

It's likely that because of the bug in #83628 this logic existed in
order to make the tests pass but it adds significant overhead depending
on the size of your `app` directory. Since #83628 matches the watcher no
longer wait for 200ms it should no longer be needed. Let's see how the
tests do with the change.