fix: isRevalidate -> isStaticGeneration in fetch cache / unstable cache (#83858)
When fetch/`unstable_cache` are stale during ISR, we leverage `workStore.isRevalidate` to determine if we're in an ISR revalidation scope and if so, we perform a blocking revalidation.

This same scenario could happen at build time if the data cache were being leveraged during build (ie with a custom cache handler, or once we enable it in the default Next cache handler). 

There's no need to distinguish between an ISR revalidation and build time prerender when checking to see if the cache entry needs to perform a blocking revalidate, so this refactors to switch to `isStaticGeneration` instead. This matches what we do in `use cache`: https://github.com/vercel/next.js/blob/1fc18244547263409a82db35edbe5908ba403410/packages/next/src/server/use-cache/use-cache-wrapper.ts#L1418