refactor: remove isRevalidate from work store (#83859)
With the refactor in #83858 there's no longer a reason to distinguish
between `isRevalidate` and `isStaticGeneration` and most likely doing so
would be a bug, as you'd likely want to apply the same handling to build
time generation that you would during a later regeneration.

This PR removes `isRevalidate` from the `WorkStore`.