[turbopack] Drop turbo_tasks::function from EcmascriptAnalyzable::module_content (#83277)
This function simply plugs two other VCs together, and it never gets a cache hit because it is "dominated" by other tasks.

The only reason it is a turbotask is because it is defined as a trait item, however, we recently (#79217) allowed for value_trait items to not be turbo tasks.  This PR builds on that to additionally allow for non-turbo_task trait items to accept `Vc<Self>`.  This only really works if there is also a default implementation which works for this case.

A tiny speed up progression in production

![image.png](https://app.graphite.dev/user-attachments/assets/c6d82929-5fb2-4921-b566-d648751a00ee.png)