[turbopack] Set `liveness` of exports based on assignment analysis (#82802)
Use the analysis added in #82763 to annotate exports as 'live' or 'const' 

This triggers previously implemented support to then bind the resulting values as 'readonly values' instead of getter functions.  This helps with both code size and execution time since parsing/allocating even these tiny getter functions is not free.