Turbopack: increase retry duration, reduce busy looping when there is other work (#84171)
### What?

The short retry timeout of 10s could lead to false panics. Increasing the retry count to 60s makes it less likely.

There is this case where the job need to wait for another job in the same queue to continue. This is handled by re-enqueuing the job at the end of the queue again. But it's not optimal to spend 1ms in busy waiting initially. So early retries will re-enqueue faster and the busy loop is only used aggressively after a bunch of retries.