docs:Small improvement on the contribution guide for clarity (#84200)
### What?
Improves the contribution guide (`01-contribution-guide.mdx`) by fixing
a typo and clarifying wording for JavaScript files with JSX code.

### Why?
This is my first time ontributing to open source, i've taken my time to
read the contribution guide and noticed this small inconsisntency. My
motivation was to start small by fixing this issue and make a
contribution, with the intention of following up with greater
improvements later. This change also helps make the documentation
clearer for new contributors and keeps it aligned with the Next.js style
guide.

### How?
**Before:**
> - Make sure to use `js` extension with JSX code at JavaScript files.

**After:**
> - Make sure to use `.js` extension for JavaScript files with **JSX**
code.

Co-authored-by: Joseph <joseph.chamochumbi@vercel.com>