Development: Skip route matching when there is an existing match (#84227)
## What?

Adds a fast path when there is an existing match already from matching
earlier in the request lifecycle. This works for the majority of cases
except for Parallel Routes as it will end up finding the wrong bundle
path. That is fundamentally related to the implementation of parallel
routes and where the bundles are created (under the specific parallel
route name). This will need a larger refactor to fix but that doesn't
have to block landing the fast path for all other cases.