simplify the way we check for turbopack config to ensure we support an empty turbopack object (#84494)
Previously we 'flattened' all the keys in the config to find unsupported leaf properties, but this meant that an empty object woudln't produce any keys.

instead just directly check for the `webpack` and `turbopack` properties and only use the scanning to look for experimental turbopack options and unsupported experimental configuration.

Fixes an issue where you cannot bypass the warning simply by adding `turbopack: {}`