Make turbopack the default bundler for custom servers (#84281)
Make Turbopack the default bundler for custom servers 🚀 

In the same line as #84216 we want to ship turbopack by default, but now for programmatic usecases.

### What?

Have the custom server entrypoint perform the same 'turbopack by default' logic as the main next entrypoints.  This was a little tricky since next itself uses the custom server entrypoints.  So i have adjusted this function to only perform the logic if it is a custom server since otherwise the next process already has the correct environment variables set

Also deprecate the `turbo` property while we are there.