make withRspack idempotent (#84482)
`loadConfig` can get called multiple times in a build or dev session.

in `dev` it gets called multiple times at the end of a session and of course whenever a session restarts
in `build` it can get called multiple times due to build workers

For that reason it needs to be idempotent, so change `withRspack` to detect if it has  already run and short circuit.  Otherwise it will always fail the second time complaining that you passed `--wepack` (because the first run deleted the TURBOPACK env var).

See discussion: https://github.com/vercel/next.js/pull/84394#issuecomment-3364615461