[segment cache]: collectSegmentData should respect experimental.staleTime config (#84520)
When we calculate the desired `staleTime` value we leverage the
`selectStaleTime` util which has some logic to fallback to the
`experimental.staleTimes` configuration if not otherwise overridden in
the prerenderStore. Without this change, the static staleTime from the
`/_tree` requests would always be the max (1 year), regardless of user
configuration.

This was caught by `prefetching.stale-times.test.ts` which I have
re-enabled & refactored in the next PR in the stack.