Allow silencing unhandled rejection filter warnings and improve debugging (#84572)
The original debug logs were added to help validate the filter worked
correctly in real applications. They leak implementation details which
is more permissive than ideal. One reason to keep these around however
is if you end up seeing the filter being removed you might want to
understand what code is actually doing the removal. The debug mode now
allows you to receive stacks so you can pinpoint what is manipulating
the listeners.

Additionally if you are ok with whatever is leading to these warnings
you might want to quiet them. This introduces a silent mode

Additionally this removes any warnings except for the one about
uninstalling the filter because doing so is so likely to be a bad idea
that we must highlight it if it is happening.