Build: Highlight TypeScript and ESLint being run clearer (#84606)
## What?

Follow-up to #84602.

- Renamed `Linting and checking validity of types` to `Running ESLint
and TypeScript concurrently`
- When `Running ESLint and TypeScript concurrently` finishes it logs the
individual times for `TypeScript` and `ESLint` separately

Previous PR only showed the time for TypeScript, hiding the time it took
to run ESlint.

The reason for renaming it to `Running ESLint and TypeScript` is that it
does not hide what is happening under the hood. Someone might assume
Next.js is doing extra work here, but it is calling the underlying tools
in the application to check linting and types. We've seen cases where
running TypeScript here is 1 minute or more, that would previously be
hidden or hard to investigate for the developer as they didn't know what
"linting and type checking" entails in this context.

```
   Creating an optimized production build ...
 ✓ Compiled successfully in 764ms
 ✓ Finished TypeScript in 968ms    
 ✓ Finished ESLint in 824ms    
 ✓ Collecting page data in 203ms    
 ✓ Generating static pages (4/4) in 318ms
 ✓ Collecting build traces in 3.3s    
 ✓ Finalizing page optimization in 3.3s    
```

---------

Co-authored-by: vercel[bot] <35613825+vercel[bot]@users.noreply.github.com>