[create-next-app]: prompt to use recommended options (#84570)
`create-next-app` by default asks a lot of questions that new users
might not be familiar with or necessarily know the answer to. This
updates the flow to start by asking if you want to use the recommended
defaults. If you already have preferences saved from previous runs, it
becomes an option to re-use those preferences. Otherwise you can proceed
with the step-by-step flow.


https://github.com/user-attachments/assets/92e815df-5484-4235-841b-eb231df8377c